Add files via upload
This repository features a Jupyter notebook showcasing my pandas practice exercises. It includes various data manipulation and analysis tasks, such as data cleaning, transformation, and visualization using pandas. The notebook contains hands-on examples and detailed explanations to demonstrate my skills and understanding of pandas for data science. It serves as both a learning resource and a reference for effective data handling techniques.